Update FrogSimulation.java
forgot to return double for part B

+++ src/main/java/FrogSimulation.java
@@ -45,7 +45,7 @@ public class FrogSimulation
 				completed++;
 			total++;
 		}
-		return completed/total;
+		return (double)completed/total;
 	}
